Fixed searching + reorganized app
