+++ run.py
@@ -1,4 +1,6 @@
-from app.app import app
+from app import create_app
 
+app = create_app()
 
-app.run(host='0.0.0.0', port=8080, debug=True)
+if __name__ == '__main__':
+    app.run(debug=True)
